split the time search filters out from the TimeSearchComponent to be used for the fliters aswell
