+++ src/api/api-client.ts
@@ -67,9 +67,18 @@ export class APIServiceClient {
 		start: string,
 		sort: string,
 		uuid: string,
+		months: string,
+		days: string,
+		timeslots: string,
 	): Promise<APISearchResponseType> {
+		months === undefined ? months == '' : null;
+		days === undefined ? months == '' : null;
+		timeslots === undefined ? months == '' : null;
+
 		return await this.httpClient.get(
-			`search/?q=${encodeURIComponent(query)}&facet=false${filters}${start}${sort}&queryUUID=${uuid}&rows=${rowCount}`,
+			`search/?q=${encodeURIComponent(
+				query,
+			)}&facet=false${filters}${start}${sort}&queryUUID=${uuid}&rows=${rowCount}${months}${days}${timeslots}`,
 		);
 	}
 
@@ -85,7 +94,7 @@ export class APIServiceClient {
 		return await this.httpClient.get(
 			`search/?q=${encodeURIComponent(
 				'*:*',
-			)}&facet=false&queryUUID=${uuid}&fq=temporal_start_year:${timeConstraint}&rows=8${months}${days}${timeslots}`,
+			)}&facet=false&queryUUID=${uuid}&fq=temporal_start_year:${timeConstraint}&rows=8${months}${days}${timeslots}&sort=startTime asc`,
 		);
 	}
