+++ src/locales/en.json
@@ -65,7 +65,7 @@
    },
   "search":{
      "searchbtn":"Search",
-     "hitcount": "{n} result fund for | {n} results found for",
+     "hitcount": " result fund for | results found for",
      "nohit":"Sorry we didn't find any results matching this search \"{query}\"",
      "showFilters":"Show filters",
      "hideFilters":"Hide filters",
